# docs/swagger.yaml
basePath: /
definitions:
  core.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  server.User:
    properties:
      avatar_url:
        type: string
      id:
        type: integer
      user_type:
        type: string
      username:
        type: string
    type: object
...
  tracker.CreateTrackerRequest:
    properties:
      body:
        type: string
      chart_config:
        type: string
      description:
        type: string
      name:
        type: string
      type:
        description: '"tracker", defaults to "tracker"'
        type: string
      visibility:
        description: 'required: "public"|"private"'
        type: string
    type: object
...
  tracker.ListTrackersResponse:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      trackers:
        items:
          $ref: '#/definitions/tracker.TrackerResponse'
        type: array
    type: object
...
  tracker.PreviewSeriesValues:
    properties:
      series:
        $ref: '#/definitions/tracker.SeriesModel'
      values:
        items:
          $ref: '#/definitions/tracker.ValueModel'
        type: array
    type: object
  tracker.SeriesModel:
    properties:
      config:
        type: string
      data_type:
        type: string
      id:
        type: integer
      name:
        type: string
      tracker_id:
        type: integer
    type: object
  tracker.TrackerModel:
    properties:
      body:
        type: string
      chart_config:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      last_updated_at:
        type: string
      name:
        type: string
      owner_id:
        type: integer
      type:
        type: string
      visibility:
        type: string
    type: object
  tracker.TrackerResponse:
    properties:
      body:
        type: string
      chart_config:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      last_updated_at:
        type: string
      like_count:
        type: integer
      liked:
        type: boolean
      name:
        type: string
      owner_id:
        type: integer
      owner_name:
        type: string
      role:
        description: '"" | "owner" | "editor"'
        type: string
      series:
        description: |-
          Series carries preview data (all series plus their latest values) when
          the list is requested with include=preview. Omitted otherwise.
        items:
          $ref: '#/definitions/tracker.PreviewSeriesValues'
        type: array
      type:
        type: string
      visibility:
        description: '"public" | "private"'
        type: string
    type: object
  tracker.ValueModel:
    properties:
      id:
        type: integer
      seriesId:
        type: integer
      time:
        type: string
      value:
        type: number
    type: object
...
info:
  contact: {}
  description: Mora is a code coverage tracker API.
  title: Mora API
  version: "1.0"
paths:
...
  /api/trackers:
    get:
      description: Return trackers owned, edited, or liked by the current user. With
        include=preview, each tracker also embeds its preview data (all series plus
        their latest values) fetched in a few batched queries.
      parameters:
      - description: Set to \
        in: query
        name: include
        type: string
      - description: Search query (partial match on tracker name)
        in: query
        name: q
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tracker.ListTrackersResponse'
      summary: List trackers for current user
      tags:
      - server
    post:
      consumes:
      - application/json
      description: Add a new tracker. The name must be unique. Creator becomes owner.
      parameters:
      - description: Tracker information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tracker.CreateTrackerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tracker.TrackerModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Create a tracker
      tags:
      - tracker
...
  /api/users/{userName}:
    get:
      description: Return the public user profile for the given username
      parameters:
      - description: Username
        in: path
        name: userName
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Get a user by username
      tags:
      - server
  /api/users/{userName}/trackers:
    get:
      description: Return trackers owned by the user. Public trackers are visible
        to everyone; private ones only to the owner.
      parameters:
      - description: Username
        in: path
        name: userName
        required: true
        type: string
      - description: Set to \
        in: query
        name: include
        type: string
      - description: Search query (partial match on tracker name)
        in: query
        name: q
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tracker.ListTrackersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: List trackers owned by a user
      tags:
      - server
...
swagger: "2.0"